% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_utils.R
\name{add_axis_text_colors}
\alias{add_axis_text_colors}
\title{Add color to axis text labels in ggplot}
\usage{
add_axis_text_colors(plot, labels, label_colors = NULL, axis = c("x", "y"))
}
\arguments{
\item{plot}{A ggplot object}

\item{labels}{Axis text labels}

\item{label_colors}{Data vector to use for coloring axis text labels}

\item{axis}{One of "x" or "y" indicating which axis.}
}
\value{
A ggplot object with the axis text labels colored.
}
\description{
Helper function to add color to axis text labels in a ggplot
  object.
}
\keyword{internal}
