% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eda_tables.R
\name{dataSummary}
\alias{dataSummary}
\title{Summary table of (X, y) data}
\usage{
dataSummary(
  X,
  y,
  digits = 2,
  sigfig = FALSE,
  features = NULL,
  max_features = 1000,
  html = knitr::is_html_output(),
  ...
)
}
\arguments{
\item{X}{Data matrix or data frame.}

\item{y}{Response vector.}

\item{digits}{Number of digits to display for numeric values}

\item{sigfig}{Logical. If \code{TRUE}, \code{digits} refers to the number of
significant figures. If \code{FALSE}, \code{digits} refers to the number of
decimal places.}

\item{features}{(Optional) vector of features to include in summary. Default
(\code{NULL}) is to include all features.}

\item{max_features}{(Optional) maximum number of features to include in 
summary. Only used if \code{features = NULL}. Default is 1000. If the 
number of features in X exceeds `max_features`, the features kept in the 
summary are chosen randomly.}

\item{html}{Logical indicating whether or not the output is an html table
or a latex table.}

\item{...}{Additional arguments to pass to pretty_DT() if
\code{html = TRUE} or pretty_kable() if \code{html = FALSE}.}
}
\value{
Returns an html table (i.e., the output of pretty_DT()) or
  a latex table (i.e., the output of pretty_kable()), 
  containing a broad overview of summary statistics for each data column.
}
\description{
Provides a summary of the given (X, y) data in table form. 
  Serves as a wrapper function around skimr::skim(), which skims a data 
  frame and returns a broad overview of useful summary statistics. This 
  wrapper can currently handle columns of type "factor", "numeric", 
  "character", "logical", "complex", "Date", and "POSIXct". All other column
  types are ignored.
}
