% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eda_tables.R
\name{dataTypes}
\alias{dataTypes}
\title{Summary table of data types in (X, y)}
\usage{
dataTypes(X, y, html = knitr::is_html_output(), ...)
}
\arguments{
\item{X}{Data matrix or data frame.}

\item{y}{Response vector.}

\item{html}{Logical indicating whether or not the output is an html table
or a latex table.}

\item{...}{Additional arguments to pass to simChef::pretty_DT() if
\code{html = TRUE} or simChef::pretty_kable() if \code{html = FALSE}.}
}
\value{
Returns an html table (i.e., the output of simChef::pretty_DT()) or
  a latex table (i.e., the output of simChef::pretty_kable()), 
  containing the frequency of each data type in the given (X, y) data.
}
\description{
Computes the frequency of each data type in the provided (X, y)
  data and returns an html or latex table with this information.
}
