% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-html.R
\name{default_mathjax}
\alias{default_mathjax}
\title{Mathjax (functions copied from rmarkdown)}
\usage{
default_mathjax()
}
\description{
Function copied from `rmdformats` since it is a non-exported
  function
}
\keyword{internal}
