% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globals.R
\docType{data}
\name{.chunk_globalenv}
\alias{.chunk_globalenv}
\title{Global environment variable to keep track of chunk indices for subchunkify}
\format{
An object of class \code{environment} of length 0.
}
\usage{
.chunk_globalenv
}
\description{
Global environment variable to keep track of chunk indices for subchunkify
}
\keyword{datasets}
