% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-and-preprocessing.R
\name{filter_cols}
\alias{filter_cols}
\alias{filter_cols_by_var}
\title{Filter out columns in data to reduce dimension.}
\usage{
filter_cols_by_var(X, min_var = NULL, max_p = NULL)
}
\arguments{
\item{X}{A data matrix or data frame.}

\item{min_var}{(Optional) minimum variance threshold. All columns with
variance lower than `min_var` are removed. If \code{NULL} (default), no
variance threshold is applied.}

\item{max_p}{(Optional) maximum number of features to keep. Only features
with the top `max_p` highest variances are kept. If \code{NULL} (default),
there is no limit on the maximum number of features to keep.}
}
\value{
A cleaned data matrix or data frame.
}
\description{
Given data X, filters out columns in X according to various
  data preprocessing/cleaning procedures. `filter_cols_by_var` reduces the
  number of features in the data by keeping those with the largest variance.
}
