% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_aesthetics}
\alias{get_aesthetics}
\title{Copy of `get_aesthetics()` function from simChef}
\usage{
get_aesthetics(...)
}
\arguments{
\item{...}{Arguments to pass to simChef:::get_aesthetics()}
}
\description{
Copy of `get_aesthetics()` function from simChef
}
\keyword{internal}
