% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_utils.R
\name{get_custom_color_palette}
\alias{get_custom_color_palette}
\title{Converts data vector to color vector}
\usage{
get_custom_color_palette(color_labels, return_palette = FALSE)
}
\arguments{
\item{color_labels}{Vector of labels to convert to colors.}

\item{return_palette}{Logical indicating whether or not to return the
color palette used to convert data to colors.}
}
\value{
If \code{return_palette = FALSE}, returns a vector of colors. If
  \code{return_palette = TRUE}, returns a list of two:
  \describe{
  \item{colors}{Vector of colors.}
  \item{palette}{Color palette used to convert data to colors.}
  }
}
\description{
Converts data vector to color vector
}
\keyword{internal}
