% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables-eda.R
\name{get_data_dimensions}
\alias{get_data_dimensions}
\title{Text summary of the data dimensions}
\usage{
get_data_dimensions(Xtrain, Xvalid, Xtest, print = TRUE)
}
\arguments{
\item{Xtrain}{Training data matrix or data frame.}

\item{Xvalid}{Validation data matrix or data frame.}

\item{Xtest}{Test data matrix or data frame.}

\item{print}{Logical indicating whether or not to print the text summary to
the console.}
}
\value{
A character string with the text summary of the data dimensions.
}
\description{
Returns a character string, summarizing the number of features 
  and samples in the training, validation, and test data sets.
}
