% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-html.R
\name{html_dependency_bootstrap}
\alias{html_dependency_bootstrap}
\title{create an html dependency for bootstrap (function copied from rmarkdown)}
\usage{
html_dependency_bootstrap(theme = "bootstrap")
}
\description{
Function copied from `rmdformats` since it is a non-exported
  function
}
\keyword{internal}
