% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/veridical.R
\name{html_dependency_bootstrap_material}
\alias{html_dependency_bootstrap_material}
\title{bootstrap material design js and css}
\usage{
html_dependency_bootstrap_material()
}
\description{
Function copied from `rmdformats` since it is a non-exported
  function. Originally from
  \url{https://github.com/FezVrasta/bootstrap-material-design}
}
\keyword{internal}
