% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vdoc.R
\name{html_dependency_material}
\alias{html_dependency_material}
\title{material js and css}
\usage{
html_dependency_material()
}
\description{
Function copied from `rmdformats` since it is a non-exported
  function.
}
\keyword{internal}
