% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-html.R
\name{html_dependency_navigation}
\alias{html_dependency_navigation}
\title{Navigation dependency}
\usage{
html_dependency_navigation(code_menu = TRUE, source_embed = FALSE)
}
\description{
Function copied from `rmdformats` since it is a non-exported
  function
}
\keyword{internal}
