% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/veridical.R
\name{html_dependency_veridical}
\alias{html_dependency_veridical}
\title{veridical dependencies}
\usage{
html_dependency_veridical(lab_notebook = FALSE)
}
\arguments{
\item{lab_notebook}{Logical indicating whether the veridical theme is being
used for a PCS lab notebook. If \code{TRUE}, the veridical theme will
also load additional scripts to enable interactivity of the notebook.}
}
\description{
Dependencies on top of base material design theme.
}
\keyword{internal}
