% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-html.R
\name{html_template}
\alias{html_template}
\title{Shared HTML template function}
\usage{
html_template(
  template_name,
  template_path,
  template_dependencies,
  pandoc_args,
  ...
)
}
\description{
Function copied from `rmdformats` since it is a non-exported
  function
}
\keyword{internal}
