% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeling-caret.R, R/modeling-h2o.R,
%   R/modeling-tidymodels.R
\name{interpret_model}
\alias{interpret_model}
\alias{interpret_caret}
\alias{interpret_h2o}
\alias{interpret_tidymodels}
\title{Wrappers for extracting feature importances from a fitted model using common
modeling backends}
\usage{
interpret_caret(fit, ...)

interpret_h2o(fit, ...)

interpret_tidymodels(fit, ...)
}
\arguments{
\item{fit}{Model fit. Typically the output of `fit_caret()`, `fit_h2o()`, or
`fit_tidymodels()`.}

\item{...}{Additional arguments to pass to `varImp()` in `fit_caret()`,
`h2o.varimp()` in `fit_h2o()`, or `vip::vi()` in `fit_tidymodels()`.}
}
\value{
A tibble with the following columns:
\describe{
\item{Variable}{Name of feature/variable.}
\item{Importance}{Feature importance score.}
}
}
\description{
`interpret_caret`, `interpret_tidymodels`, and `interpret_h2o` are
  wrappers for extracting feature importances from a fitted model using
  caret, tidymodels, and h2o backends, respectively. These wrapper functions
  provide uniformity of input arguments to easily switch between the
  different modeling packages (see `interpret_models()`).
}
\seealso{
Other interpret_models_family: 
\code{\link{interpret_models}()}
}
\concept{interpret_models_family}
