% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeling-caret.R, R/modeling-h2o.R,
%   R/modeling-tidymodels.R
\name{interpretModel}
\alias{interpretModel}
\alias{interpretCaret}
\alias{interpretH2O}
\alias{interpretTidymodels}
\title{Wrappers for extracting feature importances from a fitted model using common
modeling backends}
\usage{
interpretCaret(fit, ...)

interpretH2O(fit, ...)

interpretTidymodels(fit, ...)
}
\arguments{
\item{fit}{Model fit. Typically the output of `fitCaret()`, `fitH2O()`, or
`fitTidymodels()`.}

\item{...}{Additional arguments to pass to `varImp()` in `fitCaret()`,
`h2o.varimp()` in `fitH2O()`, or `vip::vi()` in `fitTidymodels()`.}
}
\value{
A tibble with the following columns:
\describe{
\item{Variable}{Name of feature/variable.}
\item{Importance}{Feature importance score.}
}
}
\description{
`interpretCaret`, `interpretTidymodels`, and `interpretH2O` are
  wrappers for extracting feature importances from a fitted model using
  caret, tidymodels, and h2o backends, respectively. These wrapper functions
  provide uniformity of input arguments to easily switch between the
  different modeling packages (see `interpretModels()`).
}
\seealso{
Other interpret_models_family: 
\code{\link{interpretModels}()}
}
\concept{interpret_models_family}
