% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-knitr.R
\name{load_saved_responses}
\alias{load_saved_responses}
\title{Load in saved responses in vdocs Rmd template}
\usage{
load_saved_responses()
}
\description{
Helper function to load in previously saved responses into vdocs
  Rmarkdown html report.
}
