% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{loadBRCAData}
\alias{loadBRCAData}
\title{Loads toy TCGA BRCA data}
\usage{
loadBRCAData()
}
\value{
A list of two:
\describe{
\item{X}{A data frame of gene expression features, measured via RNA-Seq.}
\item{y}{Response vector; PAM50 BRCA subtypes.}
}
}
\description{
Downloads and loads in TCGA BRCA data as a toy example for the
  PCS lab notebook
}
