% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{loadFile}
\alias{loadFile}
\title{Load in data file}
\usage{
loadFile(file_path)
}
\arguments{
\item{file_path}{Path to file to load in.}
}
\value{
Output of `readRDS()` for .rds files and `data.table::fread()` for
  all other file types.
}
\description{
Loads in files using `readRDS()` for .rds files and
  `data.table::fread()` for all other file types.
}
