% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots-ggplot-wrappers.R
\name{plot_bar}
\alias{plot_bar}
\title{Plot pretty bar plots using custom ggplot theme.}
\usage{
plot_bar(
  data,
  x_str,
  y_str = NULL,
  fill_str = NULL,
  fill = "#6FBBE3",
  stat = "count",
  position = "dodge",
  theme_options = NULL,
  show_plot = FALSE,
  ...
)
}
\arguments{
\item{data}{Data frame to use for plot.}

\item{x_str}{Character string. Name of variable in \code{data} to plot on
x-axis.}

\item{y_str}{Character string. Name of variable in \code{data} to plot on
y-axis. Used only if \code{stat = "identity"}.}

\item{fill_str}{Character string (optional). Name of variable in \code{data}
to use as fill aesthetic in plot.}

\item{fill}{Fill color. Used only if \code{fill_str} is \code{NULL}.}

\item{stat}{Override the default connection between \code{geom_bar()} and
\code{stat_count()}.}

\item{position}{Position adjustment, either as a string, or the result of
a call to a position adjustment function.}

\item{theme_options}{(Optional) list of arguments to pass to
pretty_ggplot_theme().}

\item{show_plot}{Logical. Should this plot be printed? Default \code{FALSE}.}

\item{...}{Other arguments to pass to [ggplot2::geom_bar()].}
}
\value{
A ggplot bar plot.
}
\description{
Wrapper around [ggplot2::geom_bar()] that plots pretty bar
  plots using a custom ggplot theme.
}
\examples{
df <- data.frame(x = rep(letters[1:3], 2), y = rep(LETTERS[1:2], 3))
plot_bar(data = df, x_str = "x")
plot_bar(df, x_str = "x", fill_str = "y")

}
\seealso{
Other pretty_ggplot_wrappers: 
\code{\link{plot_boxplot}()},
\code{\link{plot_density}()},
\code{\link{plot_histogram}()},
\code{\link{plot_line}()},
\code{\link{plot_point}()}
}
\concept{pretty_ggplot_wrappers}
