% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots-ggplot-wrappers.R
\name{plot_boxplot}
\alias{plot_boxplot}
\title{Plot pretty boxplots using custom ggplot theme.}
\usage{
plot_boxplot(
  data,
  x_str = NULL,
  y_str = NULL,
  fill_str = NULL,
  horizontal = FALSE,
  theme_options = NULL,
  show_plot = FALSE,
  ...
)
}
\arguments{
\item{data}{Data frame to use for plot.}

\item{x_str}{Character string. Name of variable in \code{data} to plot on
x-axis. If \code{NULL}, plot boxplot using all values in data frame.}

\item{y_str}{Character string (optional). Name of variable in \code{data} to
plot on y-axis. Should be a factor variable.}

\item{fill_str}{Character string (optional). Name of variable in \code{data}
to use as fill aesthetic in plot.}

\item{horizontal}{Logical. Whether the boxplots should be horizontal instead
of vertical.}

\item{theme_options}{(Optional) list of arguments to pass to
pretty_ggplot_theme().}

\item{show_plot}{Logical. Should this plot be printed? Default \code{FALSE}.}

\item{...}{Other arguments to pass to [ggplot2::geom_boxplot()].}
}
\value{
A ggplot boxplot.
}
\description{
Wrapper around [ggplot2::geom_boxplot()] that plots pretty
  boxplots using a custom ggplot theme.
}
\examples{
## plot boxplot of all data in data frame
plot_boxplot(as.data.frame(matrix(rnorm(1000), nrow = 100)))
## plot boxplot of single column in data frame
plot_boxplot(iris, y_str = "Sepal.Width")
plot_boxplot(iris, x_str = "Species", y_str = "Sepal.Width")
iris2 <- data.frame(iris,
                    z = as.factor(rep(letters[1:2],
                                  length.out = nrow(iris))))
plot_boxplot(iris2, x_str = "Species", y_str = "Sepal.Width", fill_str = "z")
plot_boxplot(iris2, y_str = "Sepal.Width", fill_str = "z")

}
\seealso{
Other pretty_ggplot_wrappers: 
\code{\link{plot_bar}()},
\code{\link{plot_density}()},
\code{\link{plot_histogram}()},
\code{\link{plot_line}()},
\code{\link{plot_point}()}
}
\concept{pretty_ggplot_wrappers}
