% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots-eda.R
\name{plot_data_heatmap}
\alias{plot_data_heatmap}
\title{Plots heatmap of (X, y) data}
\usage{
plot_data_heatmap(
  X,
  y,
  subsample_rows = 1,
  subsample_cols = 1,
  clust_rows = TRUE,
  clust_cols = TRUE,
  ...
)
}
\arguments{
\item{X}{Data matrix or data frame to use for heatmap.}

\item{y}{A response vector.}

\item{subsample_rows}{Proportion of rows to subsample and keep in plotting
display. Default is 1, which does not perform any subsampling.}

\item{subsample_cols}{Proportion of columns to subsample and keep in plotting
display. Default is 1, which does not perform any subsampling.}

\item{clust_rows}{Logical indicating whether or not to cluster the rows in X
via hierarchical clustering.}

\item{clust_cols}{Logical indicating whether or not to cluster the columns
in X via hierarchical clustering.}

\item{...}{Additional arguments to pass to plot_hclust_heatmap()}
}
\value{
A gglot object. Specifically, if y is a continuous response vector,
  then this function returns a `patchwork` object with two ggplots: one for
  the heatmap of X and another for the scatter plot of y values. If y is
  categorical, this function returns a single ggplot object with the
  heatmap of X with samples grouped by each category in y.
}
\description{
Returns a heatmap of the X data alongside the response values y.
  Here, the fill colors in the heatmap correspond to the values in X.
  If the response vector y is categorical, the rows/samples in X are grouped
  into their respective y categories. If the response vector is continuous,
  a scatter plot of the y values are plotted to the right of the heatmap.
  Furthermore, the rows and columns of X can be clustered via hierarchical
  clustering if desired.
}
