% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots-ggplot-wrappers.R
\name{plot_density}
\alias{plot_density}
\title{Plot pretty kernel density plots using custom ggplot theme.}
\usage{
plot_density(
  data,
  x_str = NULL,
  fill_str = NULL,
  fill = "#6FBBE3",
  alpha = 0.4,
  theme_options = NULL,
  show_plot = FALSE,
  ...
)
}
\arguments{
\item{data}{Data frame to use for plot.}

\item{x_str}{Character string. Name of variable in \code{data} to plot on
x-axis. If \code{NULL}, plot density using all values in data frame.}

\item{fill_str}{Character string (optional). Name of variable in \code{data}
to use as fill aesthetic in plot.}

\item{fill}{Fill color. Used only if \code{fill_str} is \code{NULL}.}

\item{alpha}{Alpha value for transparency.}

\item{theme_options}{(Optional) list of arguments to pass to
vthemes::theme_vmodern().}

\item{show_plot}{Logical. Should this plot be printed? Default \code{FALSE}.}

\item{...}{Other arguments to pass to [ggplot2::geom_density()].}
}
\value{
A ggplot density.
}
\description{
Wrapper around [ggplot2::geom_density()] that plots pretty
  density plots using a custom ggplot theme.
}
\examples{
## plot distribution of all data in data frame
plot_density(as.data.frame(rnorm(1000), nrow = 100))
## plot distribution of a single column in data frame
plot_density(iris, x_str = "Sepal.Width")
plot_density(iris, x_str = "Sepal.Width", fill_str = "Species")

}
\seealso{
Other pretty_ggplot_wrappers: 
\code{\link{plot_bar}()},
\code{\link{plot_boxplot}()},
\code{\link{plot_histogram}()},
\code{\link{plot_line}()},
\code{\link{plot_point}()}
}
\concept{pretty_ggplot_wrappers}
