% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots-ggplot-wrappers.R
\name{plot_line}
\alias{plot_line}
\title{Plot pretty line plots using custom ggplot theme.}
\usage{
plot_line(
  data,
  x_str,
  y_str,
  color_str = NULL,
  theme_options = NULL,
  show_plot = FALSE,
  ...
)
}
\arguments{
\item{data}{Data frame to use for plot.}

\item{x_str}{Character string. Name of variable in \code{data} to plot on
x-axis.}

\item{y_str}{Character string. Name of variable in \code{data} to plot on
y-axis.}

\item{color_str}{Character string (optional). Name of variable in \code{data}
to use as color aesthetic in plot.}

\item{theme_options}{(Optional) list of arguments to pass to
pretty_ggplot_theme().}

\item{show_plot}{Logical. Should this plot be printed? Default \code{FALSE}.}

\item{...}{Other arguments to pass to [ggplot2::geom_line()].}
}
\value{
A ggplot line plot.
}
\description{
Wrapper around [ggplot2::geom_line()] that plots pretty
  line plots using a custom ggplot theme.
}
\examples{
df <- data.frame(time = 1:5, value = 5:9)
plot_line(df, x_str = "time", y_str = "value")
df2 <- data.frame(time = rep(1:5, 2), value = 1:10,
                  group = rep(letters[1:2], each = 5))
plot_line(df2, x_str = "time", y_str = "value", color_str = "group")

}
\seealso{
Other pretty_ggplot_wrappers: 
\code{\link{plot_bar}()},
\code{\link{plot_boxplot}()},
\code{\link{plot_density}()},
\code{\link{plot_histogram}()},
\code{\link{plot_point}()}
}
\concept{pretty_ggplot_wrappers}
