% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots-ggplot-wrappers.R
\name{plot_point}
\alias{plot_point}
\title{Plot pretty scatter plots using custom ggplot theme.}
\usage{
plot_point(
  data,
  x_str,
  y_str,
  color_str = NULL,
  theme_options = NULL,
  show_plot = FALSE,
  ...
)
}
\arguments{
\item{data}{Data frame to use for plot.}

\item{x_str}{Character string. Name of variable in \code{data} to plot on
x-axis.}

\item{y_str}{Character string. Name of variable in \code{data} to plot on
y-axis.}

\item{color_str}{Character string (optional). Name of variable in \code{data}
to use as color aesthetic in plot.}

\item{theme_options}{(Optional) list of arguments to pass to
vthemes::theme_vmodern().}

\item{show_plot}{Logical. Should this plot be printed? Default \code{FALSE}.}

\item{...}{Other arguments to pass to [ggplot2::geom_point()].}
}
\value{
A ggplot scatter plot.
}
\description{
Wrapper around [ggplot2::geom_point()] that plots pretty
  scatter plots using a custom ggplot theme.
}
\examples{
plot_point(iris, x_str = "Sepal.Width", y_str = "Sepal.Length")
plot_point(iris, x_str = "Sepal.Width", y_str = "Sepal.Length",
            color_str = "Species")

}
\seealso{
Other pretty_ggplot_wrappers: 
\code{\link{plot_bar}()},
\code{\link{plot_boxplot}()},
\code{\link{plot_density}()},
\code{\link{plot_histogram}()},
\code{\link{plot_line}()}
}
\concept{pretty_ggplot_wrappers}
