% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eda_plots.R
\name{plotDataDistribution}
\alias{plotDataDistribution}
\title{Plots summary of data distribution}
\usage{
plotDataDistribution(
  data,
  by_feature = NULL,
  plot_type = "auto",
  xlab = "Value",
  title = NULL,
  plot_heights = 1,
  theme_options = NULL,
  ...
)
}
\arguments{
\item{data}{A data matrix, data frame, or vector.}

\item{by_feature}{Logical. If \code{TRUE}, plots distributions for each
feature separately. If \code{FALSE}, plots distribution of all features
together. Default is \code{TRUE} if there are <10 features and \code{FALSE}
otherwise.}

\item{plot_type}{Type of plot. Default is "auto", which uses a kernel
density plot for continuous features and a bar plot for categorical
features. If not "auto", `plot_type` should be a list with two named
elements: `continuous` and `categorical`. The `continuous` element must be
one of "density", "histogram", and "boxplot" while the `categorical`
element must be "bar" (with more options to come), indicating the type of
plot to use for continuous and categorical features, respectively.}

\item{xlab}{X-axis label.}

\item{title}{Plot title.}

\item{plot_heights}{(Optional) numeric vector of relative row heights of
subplots. Only used if both continuous and categorical features are found
in the data. For example, heights = c(2, 1) would make the first row twice
as tall as the second row.}

\item{theme_options}{(Optional) list of arguments to pass to
pretty_ggplot_theme().}

\item{...}{Additional arguments to pass to ggplot2::geom_*().}
}
\value{
A ggplot object.
}
\description{
Plots a summary of the feature distributions (either together or
  separately per feature) in the data. Only continuous (i.e., numeric) and
  categorical (i.e., character or factor) features are used for plotting.
}
