% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpretability_plots.R
\name{plotFeatureImportance}
\alias{plotFeatureImportance}
\alias{plotFeatureImportancePair}
\alias{plotFeatureImportanceStability}
\title{Feature importance plots}
\usage{
plotFeatureImportance(
  data,
  feature_col = "Variable",
  importance_col = "Importance",
  model_col = "Method",
  max_features = 50,
  use_rankings = FALSE,
  use_facets = TRUE,
  interactive = FALSE,
  ...
)

plotFeatureImportancePair(
  data,
  feature_col = "Variable",
  importance_col = "Importance",
  model_col = "Method",
  use_rankings = FALSE,
  interactive = FALSE,
  ...
)

plotFeatureImportanceStability(
  data,
  feature_col = "Variable",
  importance_col = "Importance",
  model_col = "Method",
  max_features = 50,
  use_rankings = FALSE,
  use_facets = TRUE,
  interactive = FALSE,
  ...
)
}
\arguments{
\item{data}{A data frame with the feature importance information. At a
minimum, this data frame should contain columns with the names specified by
`feature_col`, `importance_col`, and `model_col`.}

\item{feature_col}{Character string. Name of the column in `data` with the
feature names.}

\item{importance_col}{Character string. Name of the column in `data` with the
feature importance scores.}

\item{model_col}{Character string. Name of the column in `data` with the
model names.}

\item{max_features}{Maximum number of features to display in the plot.}

\item{use_rankings}{Logical indicating whether or not to use the feature
importance rankings, rather than the raw feature importance scores, for
plotting.}

\item{use_facets}{Logical indicating whether or not to use faceting to
separate the feature importances by model.}

\item{interactive}{Logical indicating whether or not to return an interactive
plot.}

\item{...}{Additional arguments to pass to `plotPairs()`. Only used in
`plotFeatureImportancePair()`.}
}
\value{
If `interactive = FALSE`, returns a ggplot object. If
  `interactive = TRUE`, returns an interactive plotly object.
}
\description{
Various plotting functions to visualize feature importances.
  `plotFeatureImportances` creates bar plots of the feature importances
  across different models. `plotFeatureImportancePair` create pair plots,
  comparing the feature importances between all pairs of models.
  `plotFeatureImportanceStability` creates boxplots of the
  distribution of feature importances across the many data perturbations for
  each model.
}
