% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eda_plots.R
\name{plotPairs}
\alias{plotPairs}
\title{Plot pretty pair plots using custom ggplot theme.}
\usage{
plotPairs(
  data,
  columns,
  color = NULL,
  color_upper = NULL,
  color_label = "",
  color_upper_label = "",
  color_scheme = NULL,
  color_scheme_upper = NULL,
  column_labels = NULL,
  title = "",
  point_size = 0.5,
  point_alpha = 0.5,
  cor_text_size = 3.5,
  subsample = 1,
  show_upper = TRUE,
  drop = FALSE,
  theme_function = NULL,
  show_plot = FALSE,
  ...
)
}
\arguments{
\item{data}{Data frame to use for plot.}

\item{columns}{Vector of column indicies or column names to plot.}

\item{color}{(Optional) Data vector to use as colors for lower ggplot
panels.}

\item{color_upper}{(Optional) Data vector to use as colors for upper ggplot
panels.}

\item{color_label}{Character string. Label for color legend title (used
in lower ggplot panels).}

\item{color_upper_label}{Character string for color_upper legend title (used
in upper ggplot panels).}

\item{color_scheme}{(Optional) Vector of colors to set manual color
scheme corresponding to color_lower argument (i.e., the color scheme in the
lower panels). If \code{NULL} (default), viridis color scheme is used.}

\item{color_scheme_upper}{(Optional) Vector of colors to set manual color
scheme corresponding to color_upper argument (i.e., the color scheme in the
upper panels). If \code{NULL} (default), viridis color scheme is used.}

\item{column_labels}{Label names to be displayed on strips.}

\item{title}{Character string. Title of plot.}

\item{point_size}{Point size for [ggplot2::geom_point()].}

\item{point_alpha}{Alpha value for [ggplot2::geom_point()].}

\item{cor_text_size}{Size of correlation text.}

\item{subsample}{Proportion of rows to subsample and plot.}

\item{show_upper}{Logical. Should we show subplots in upper panels?}

\item{drop}{Logical. Whether or not to drop factors with no observations.}

\item{theme_function}{function which adds theme() to ggpairs() object. If
\code{NULL}, add \code{pretty_ggplot_theme()} to
[GGally::ggpairs()] object.}

\item{show_plot}{Logical. Should this plot be printed? Default \code{FALSE}.}

\item{...}{Other arguments to pass to \code{pretty_ggplot_theme()}
or \code{theme_function()}}
}
\value{
A [GGally::ggpairs] object.
}
\description{
Plots pretty pair plots using a custom ggplot theme and provides
  additional functionality beyond [GGally::ggpairs()] such as the ability to
  add two different color schemes (one for the lower triangular subpanels and
  another for the upper triangular subpanels).
}
\examples{
plotPairs(data = iris, columns = 1:ncol(iris), color = iris$Species)

}
