% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eda_plots.R
\name{plotPCA}
\alias{plotPCA}
\title{Plot pretty PCA plots using custom ggplot theme.}
\usage{
plotPCA(
  X,
  pca_obj,
  npcs,
  pcs,
  color = NULL,
  color_upper = NULL,
  color_label = "",
  color_upper_label = "",
  color_scheme = NULL,
  color_scheme_upper = NULL,
  point_size = 0.5,
  point_alpha = 1,
  subsample = 1,
  show_var = TRUE,
  center = TRUE,
  scale = FALSE,
  title = "",
  show_plot = FALSE,
  ...
)
}
\arguments{
\item{X}{Data matrix or data.frame on which to perform PCA. Must specify
either \code{X} or \code{pca_obj}.}

\item{pca_obj}{Output of previous run of plotPCA() to avoid re-computing SVDs
(i.e., the PC loadings and scores) again. Must specify either \code{X} or
\code{pca_obj}. Ignored if \code{X} is provided.}

\item{npcs}{Number of top PCs to plot. Must specify either \code{npcs} or
\code{pcs}.}

\item{pcs}{Vector of which PCs to show. Must specify either \code{npcs} or
\code{pcs}. Ignored if \code{npcs} is provided.}

\item{color}{(Optional) Data vector to use as colors for lower ggplot
panels.}

\item{color_upper}{(Optional) Data vector to use as colors for upper ggplot
panels.}

\item{color_label}{Character string. Label for color legend title (used
in lower ggplot panels).}

\item{color_upper_label}{Character string for color_upper legend title (used
in upper ggplot panels).}

\item{color_scheme}{(Optional) Vector of colors to set manual color
scheme corresponding to color_lower argument (i.e., the color scheme in the
lower panels). If \code{NULL} (default), viridis color scheme is used.}

\item{color_scheme_upper}{(Optional) Vector of colors to set manual color
scheme corresponding to color_upper argument (i.e., the color scheme in the
upper panels). If \code{NULL} (default), viridis color scheme is used.}

\item{point_size}{Point size for [ggplot2::geom_point()].}

\item{point_alpha}{Alpha value for [ggplot2::geom_point()].}

\item{subsample}{Proportion of rows to subsample and plot.}

\item{show_var}{Logical. Whether or not to show the proportion of variance
explained in axes labels.}

\item{center}{Logical. Whether or not to center data for PCA.}

\item{scale}{Logical. Whether or not to scale data for PCA.}

\item{title}{Character string. Title of plot.}

\item{show_plot}{Logical. Should this plot be printed? Default \code{FALSE}.}

\item{...}{Other arguments to pass to \code{pretty_ggplot_theme()}
or \code{theme_function()}}
}
\value{
A list of four:
\describe{
\item{plot}{A ggplot object of the PC pair plots.}
\item{scores}{A matrix with the PC scores.}
\item{loadings}{A matrix with the PC loadings.}
\item{var.explained}{A vector of the proportions of variance explained.}
}
}
\description{
Run PCA on the given data matrix and generates PC plots for the
  specified principal components.
}
\examples{
out <- plotPCA(X = iris[, -5], npcs = 3, color = iris$Species)
out$plot
iris2 <- data.frame(iris, z = rep(letters[1:2], length.out = nrow(iris)))
out <- plotPCA(X = iris2[, -c(5, 6)], npcs = 3,
               color = iris2$Species, color_upper = as.factor(iris2$z))
out$plot

}
