% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeling-caret.R, R/modeling-h2o.R,
%   R/modeling-tidymodels.R
\name{predict_model}
\alias{predict_model}
\alias{predict_caret}
\alias{predict_h2o}
\alias{predict_tidymodels}
\title{Wrappers for making predictions from a fitted model using common modeling
backends}
\usage{
predict_caret(fit, Xtest, ...)

predict_h2o(fit, Xtest, ...)

predict_tidymodels(fit, Xtest, ...)
}
\arguments{
\item{fit}{Model fit. Typically the output of `fit_caret()`, `fit_h2o()`, or
`fit_tidymodels()`.}

\item{Xtest}{Data matrix or data frame on which to make predictions.}

\item{...}{Additional arguments to pass to `predict.train()` in `fit_caret()`,
`h2o.predict()` in `fit_h2o()`, or `predict.workflow()` in
`fit_tidymodels()`.}
}
\value{
A tibble with the following columns:
\describe{
\item{predictions}{Raw predicted value (e.g., the predicted class in a
  classification problem and the predicted continuous value in a
  regression problem.)}
\item{prob_predictions}{In a classification problem, this is a tibble with
  the predicted probabilities for each class. This column is omitted for
  regression problems.}
}
}
\description{
`predict_caret`, `predict_tidymodels`, and `predict_h2o` are
  wrappers for making predictions from a model using caret, tidymodels, and
  h2o backends, respectively. These wrapper functions provide uniformity of
  input arguments to easily switch between the different modeling packages
  (see `predict_models()`).
}
\seealso{
Other predict_models_family: 
\code{\link{predict_models}()}
}
\concept{predict_models_family}
