% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeling-caret.R, R/modeling-h2o.R,
%   R/modeling-tidymodels.R
\name{predictModel}
\alias{predictModel}
\alias{predictCaret}
\alias{predictH2O}
\alias{predictTidymodels}
\title{Wrappers for making predictions from a fitted model using common modeling
backends}
\usage{
predictCaret(fit, Xtest, ...)

predictH2O(fit, Xtest, ...)

predictTidymodels(fit, Xtest, ...)
}
\arguments{
\item{fit}{Model fit. Typically the output of `fitCaret()`, `fitH2O()`, or
`fitTidymodels()`.}

\item{Xtest}{Data matrix or data frame on which to make predictions.}

\item{...}{Additional arguments to pass to `predict.train()` in `fitCaret()`,
`h2o.predict()` in `fitH2O()`, or `predict.workflow()` in
`fitTidymodels()`.}
}
\value{
A tibble with the following columns:
\describe{
\item{predictions}{Raw predicted value (e.g., the predicted class in a
  classification problem and the predicted continuous value in a
  regression problem.)}
\item{prob_predictions}{In a classification problem, this is a tibble with
  the predicted probabilities for each class. This column is omitted for
  regression problems.}
}
}
\description{
`predictCaret`, `predictTidymodels`, and `predictH2O` are
  wrappers for making predictions from a model using caret, tidymodels, and
  h2o backends, respectively. These wrapper functions provide uniformity of
  input arguments to easily switch between the different modeling packages
  (see `predictModels()`).
}
\seealso{
Other predict_models_family: 
\code{\link{predictModels}()}
}
\concept{predict_models_family}
