% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeling.R
\name{predictModels}
\alias{predictModels}
\title{Wrapper for making predictions from fitted models using common modeling
backends}
\usage{
predictModels(fit_list, Xtest, options = list(), use = NULL)
}
\arguments{
\item{fit_list}{List of fitted models for which to make predictions.
Typically the output of `fitModels()`.}

\item{Xtest}{Data matrix or data frame on which to make predictions.}

\item{options}{Named list of additional arguments to pass to
`predict.train()` if using caret backend, `h2o.predict()` if using h2o
backend, or `predict.workflow()` if using tidymodels backend. If different
arguments need to be passed in for different fitted models, `options`
should be a list where each element is named after the model (i.e., the
same name as in `fit_list`) and is a list of named arguments. If `options`
is not a list of list, then all list elements are passed to the predict
function for every model.}

\item{use}{One of "caret", "h2o", "tidymodels", indicating the modeling
package to use. If \code{NULL} (default), the modeling package is inferred
from the fit.}
}
\value{
A tibble with the following columns:
\describe{
\item{Method}{Name of method that made the associated prediction.}
\item{predictions}{Raw predicted value (e.g., the predicted class in a
  classification problem and the predicted continuous value in a
  regression problem.)}
\item{prob_predictions}{In a classification problem, this is a tibble with
  the predicted probabilities for each class. This column is omitted for
  regression problems.}
}

Note that the order of the predictions within each method are aligned with
that of the Xtest input.
}
\description{
`predictModels` is a wrapper function for making predictions
  from multiple fitted models that were trained using caret, tidymodels, or
  h2o backends. This wrapper function provides uniformity of input arguments
  to easily switch between the different modeling packages.
}
\seealso{
Other predict_models_family: 
\code{\link{predictModel}}
}
\concept{predict_models_family}
