% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeling-caret.R, R/modeling-h2o.R,
%   R/modeling-tidymodels.R
\name{print_fit}
\alias{print_fit}
\alias{print_caret_fit}
\alias{print_h2o_fit}
\alias{print_tidymodels_fit}
\title{Wrappers for printing model fits from common modeling backends}
\usage{
print_caret_fit(fit)

print_h2o_fit(fit)

print_tidymodels_fit(fit)
}
\arguments{
\item{fit}{Model fit. Typically the output of `fit_caret()`, `fit_h2o()`, or
`fit_tidymodels()`.}
}
\value{
Print fit summary to the console.
}
\description{
`print_caret_fit`, `print_tidymodels_fit`, and `print_h2o_fit` are
  wrappers for printing a summary of the model fit from models that were
  trained using caret, tidymodels, and h2o backends, respectively. These
  wrapper functions provide uniformity of input arguments to easily switch
  between the different modeling packages (see `print_fit_results()`).
}
\seealso{
Other print_fit_family: 
\code{\link{print_fit_results}()}
}
\concept{print_fit_family}
