% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeling.R
\name{printFitResults}
\alias{printFitResults}
\title{Print fit results summary}
\usage{
printFitResults(fit_list, use = NULL, html = knitr::is_html_output())
}
\arguments{
\item{fit_list}{List of fitted models for which to make predictions.
Typically the output of `fitModels()`.}

\item{use}{One of "caret", "h2o", "tidymodels", indicating the modeling
package to use. If \code{NULL} (default), the modeling package is inferred
from the fit.}

\item{html}{Logical indicating whether the output should be in html format.
If \code{FALSE}, output is in latex format.}
}
\value{
Print fit summary to the console.
}
\description{
`printFitResults` is a wrapper function for printing a summary
  of the model fit from models that were trained using caret,
  tidymodels, or h2o backends. This wrapper function provides uniformity of
  input arguments to easily switch between the different modeling packages.
}
\seealso{
Other print_fit_family: 
\code{\link{printFit}}
}
\concept{print_fit_family}
