% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-and-preprocessing.R
\name{remove_cols}
\alias{remove_cols}
\alias{remove_na_cols}
\alias{remove_constant_cols}
\alias{remove_duplicate_cols}
\title{Basic cleaning functions to remove columns in data.}
\usage{
remove_na_cols(X, verbose = 0)

remove_constant_cols(X, verbose = 0)

remove_duplicate_cols(X, verbose = 0)
}
\arguments{
\item{X}{A data matrix or data frame.}

\item{verbose}{Integer (0-2) indicating the level of written output.}
}
\value{
A cleaned data matrix or data frame.
}
\description{
Given data X, removes columns in X according to various
  data preprocessing/cleaning procedures. `remove_na_cols` removes all
  columns in the data with at least one NA value. `remove_constant_cols`
  removes all columns in the data that are a constant value (ignoring NAs).
  `remove_duplicate_cols` removes columns in the data that are duplicates of
  another column in the data so that each column in the resulting cleaned
  data is unique.
}
