% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-and-preprocessing.R
\name{removeCols}
\alias{removeCols}
\alias{removeNACols}
\alias{removeConstantCols}
\alias{removeDuplicateCols}
\title{Basic cleaning functions to remove columns in data.}
\usage{
removeNACols(X, verbose = 0)

removeConstantCols(X, verbose = 0)

removeDuplicateCols(X, verbose = 0)
}
\arguments{
\item{X}{A data matrix or data frame.}

\item{verbose}{Integer (0-2) indicating the level of written output.}
}
\value{
A cleaned data matrix or data frame.
}
\description{
Given data X, removes columns in X according to various
  data preprocessing/cleaning procedures. `removeNACols` removes all
  columns in the data with at least one NA value. `removeConstantCols`
  removes all columns in the data that are a constant value (ignoring NAs).
  `removeDuplicateCols` removes columns in the data that are duplicates of
  another column in the data so that each column in the resulting cleaned
  data is unique.
}
