% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeling.R
\name{showEvalResults}
\alias{showEvalResults}
\title{Display evaluation results summary in Rmd}
\usage{
showEvalResults(
  eval_results,
  test_set = FALSE,
  html = knitr::is_html_output(),
  digits = 2,
  sigfig = TRUE,
  na_disp = "NA",
  html_options = NULL,
  latex_options = NULL,
  fig_height = 6,
  fig_width = 10
)
}
\arguments{
\item{eval_results}{Output of `evaluateModels()`}

\item{test_set}{Logical indicating whether or not this evaluation is for the
test set. If `FALSE`, output is assumed to be associated with validation
set.}

\item{html}{Logical indicating whether the output should be in html format.
If \code{FALSE}, output is in latex format.}

\item{digits}{Number of digits to display for numeric values}

\item{sigfig}{Logical. If \code{TRUE}, \code{digits} refers to the number of
significant figures. If \code{FALSE}, \code{digits} refers to the number of
decimal places.}

\item{na_disp}{Character string to display if NA entry is found in table.}

\item{html_options}{List of additional named arguments to pass to 
`simChef::pretty_DT()`.}

\item{latex_options}{List of additional named arguments to pass to 
`simChef::pretty_kable()`.}

\item{fig_height}{Height of plot (in inches). See `fig.height` in the knitr
chunk options.}

\item{fig_width}{Width of plot (in inches). See `fig.width` in the knitr
chunk options.}
}
\value{
Outputs html or latex text for the R Markdown document.
}
\description{
`showEvalResults` transforms the evaluation results from
  `evaluateModels()` for output in an R Markdown document.
}
