% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{subchunkify}
\alias{subchunkify}
\title{Copy of `subchunkify()` function from simChef}
\usage{
subchunkify(...)
}
\arguments{
\item{...}{Arguments to pass to simChef:::subchunkify()}
}
\description{
Copy of `subchunkify()` function from simChef
}
