% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{validate_data_split}
\alias{validate_data_split}
\title{Validate inputs for data splitting proportions}
\usage{
validate_data_split(train_prop, valid_prop, test_prop)
}
\arguments{
\item{train_prop}{Proportion of data to put in training set}

\item{valid_prop}{Proportion of data to put in validation set}

\item{test_prop}{Proportion of data to put in test set}
}
\description{
Checks whether data splitting proportion are valid (i.e.,
  are non-negative, less than 1, and sum to 1).
}
