% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{validateData}
\alias{validateData}
\title{Validate data inputs for (X, y)}
\usage{
validateData(X, y)
}
\arguments{
\item{X}{A data matrix or data frame.}

\item{y}{A response vector.}
}
\description{
Checks whether data dimensions for X and y match.
}
