% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-knitr.R
\name{vdoc_knitr_setup}
\alias{vdoc_knitr_setup}
\title{Knitr options setup for vdoc Rmd template}
\usage{
vdoc_knitr_setup(
  echo = TRUE,
  warning = FALSE,
  message = FALSE,
  cache = FALSE,
  collapse = TRUE,
  fig.align = "center",
  fig.pos = "H",
  fig.show = "hold",
  comment = "#>",
  ...
)
}
\arguments{
\item{echo}{Logical indicating whether to display the source code in the
output document.}

\item{warning}{Logical indicating whether to display warnings in the
output document.}

\item{message}{Logical indicating whether to display messages in the
output document.}

\item{cache}{Logical indicating whether to cache code chunks.}

\item{collapse}{Logical indicating whether to collapse all the source and
output blocks from one code chunk into a single block.}

\item{fig.align}{Alignment of figures in the output document. Possible values
are "default", "left", "right", and "center".}

\item{fig.pos}{Character string for the figure position arrangement to be
used in `\\begin{figure}[]`.}

\item{fig.show}{How to show/arrange the plots. Possible values are "asis",
"hold", "animate", and "hide".}

\item{comment}{Character string. Prefix to be added before each line of the
text output.}

\item{...}{Additional knitr options to set via `knitr::opts_chunk$set()`.
See \url{https://yihui.org/knitr/options/} for possible options.}
}
\description{
Helper function to set up knitr options and hooks for the
  vdoc Rmarkdown template.
}
